Fix gascalibration cmd name
